refactor(context-brief): keep relocation selection storage-neutral
