+++ src/code_graph/citation_primitives.ts
@@ -42,6 +42,19 @@ export interface CodeGraphCitationFileRelocationFallbackV1 {
   readonly path: string;
 }
 
+/** Select eager hashes plus fallbacks whose original path is absent. */
+export function selectCodeGraphCitationContentHashTargets(
+  eagerContentHashes: readonly string[],
+  fileRelocationFallbacks: readonly CodeGraphCitationFileRelocationFallbackV1[],
+  presentPaths: ReadonlySet<string>,
+): readonly string[] {
+  const requestedContentHashes = new Set(eagerContentHashes);
+  for (const fallback of fileRelocationFallbacks) {
+    if (!presentPaths.has(fallback.path)) requestedContentHashes.add(fallback.contentHash);
+  }
+  return [...requestedContentHashes];
+}
+
 export interface CodeGraphEffectiveSnapshotCitationEvidenceRequest {
   readonly contentHashes?: readonly string[];
   /**
